feat: add set user config
