+++ packages/rax-plugin-app/package.json
@@ -33,6 +33,7 @@
     "less": "^3.9.0",
     "less-loader": "^5.0.0",
     "loader-utils": "^1.1.0",
+    "lodash": "^4.17.15",
     "mini-css-extract-plugin": "^0.5.0",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
     "postcss": "^7.0.17",
