+++ packages/rax-plugin-app/src/config/user/keys/devPublicPath.js
@@ -1,0 +1,7 @@
+module.exports = (config, context, value) => {
+  const { command } = context;
+
+  if (command === 'dev') {
+    config.output.publicPath(value);
+  }
+};
