+++ packages/rax-plugin-app/src/config/user/keys/outputDir.js
@@ -1,0 +1,7 @@
+const path = require('path');
+
+module.exports = (config, context, value) => {
+  const { rootDir } = context;
+
+  config.output.path(path.resolve(rootDir, value));
+};
