+++ packages/rax-plugin-app/src/config/user/keys/publicPath.js
@@ -1,0 +1,7 @@
+module.exports = (config, context, value) => {
+  const { command } = context;
+
+  if (command === 'build') {
+    config.output.publicPath(value);
+  }
+};
